+++ src/Sprite.cpp
@@ -89,6 +89,6 @@ Sprite::Sprite(string id) : ImageImpl() {
 }
 
 
-void Sprite::render(ViewportRenderer* viewport, uint16_t x, uint16_t y) {
+void Sprite::render(ViewportRenderer* viewport, uint32_t x, uint32_t y) {
 	viewport->drawImage(this, 0, 0, this->width, this->height, x, y);
 }
